requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
numpy==1.24.3
opencv-python==4.8.0.76
pillow==10.0.0
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.0
scikit-learn==1.3.0
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.7
